+++ src/Resources/Wacc/Stock/Stock.ts
@@ -1,14 +1,19 @@
 import AbstractResource from '../../AbstractResource';
 import { HttpClientConfig, HttpClientResponse } from '../../../Interfaces/HttpClient';
 
-export default class Toybox extends AbstractResource {
+export default class Stock extends AbstractResource {
     initialise(): void {
-        this.resourcePath = '/stock/toybox';
+        this.resourcePath = '/stock';
         this.defaultAccess = 'wacc';
     }
 
     public update: (...args: HttpClientConfig) => HttpClientResponse = this.createMethodFromPartialSpec({
         method: 'PATCH',
-        path: '/',
+        path: '/toybox',
+    });
+
+    public reassign: (...args: HttpClientConfig) => HttpClientResponse = this.createMethodFromPartialSpec({
+        method: 'PATCH',
+        path: '/reassign-status',
     });
 }